# server/swagger.yaml
openapi: 3.0.0
info:
  title: OneUp API
  description: API contract for the OneUp project.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /points/{id}:
    get:
      summary: Get specific points entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Points entry details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  value:
                    type: integer
                    example: 10
                  user_id:
                    type: integer
                    example: 1
                  action_id:
                    type: integer
                    example: 2
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    put:
      summary: Update specific points entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                value:
                  type: integer
                  example: 20
                action_id:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Points entry updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  value:
                    type: integer
                    example: 20
                  user_id:
                    type: integer
                    example: 1
                  action_id:
                    type: integer
                    example: 2
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    delete:
      summary: Delete specific points entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '204':
          description: Points entry deleted successfully
...